drop_old_data: Add new required param --allowed-interval
The script we invoke has a new required parameter, --allowed-interval,
which states that data to be dropped must be within this many days of
the end of the period, otherwise the drop will fail. This is intended
as a safeguard against accidentally deleting data that was supposed
to stick around.

Default the value to 7 days and update the checksums for the new
argument.

Change-Id: Ida6c6b883a2644ecaecec057d5d50318bb154f1e