drop-snapshots: Remove directory handling
This script was initially written to handle hive external
tables, where directories are manually managed. We don't
use those and instead allow hive to manage the directories
for us. The end result is that the `hdfs dfs -rm ...` command
fails due to paths not existing, as hive has already deleted
the appropriate directories. To resole, remove directory
handling.

Bug: T303831
Change-Id: Ia8f100fa428d9a3690e920a8cdf860dfe8285254