airflow: Set mjolnir max_active_runs to 1
When re-enabling the mjolnir dag after it was disabled for
awhile on 11-29 airflow created dag runs for both 11-15 and
11-22. There is no need for multiple mjolnir runs going at the
same time, they will use the same sore data and collect the
same live data from the indices. Limit max_active_runs to 1
so in the future it will only start the most recent dagrun.

Change-Id: If12ddbb2e2ac56c57c429be218d92e7570e132f6