import_cirrus_index: Use spark3
There is a bug in the version of hive used by spark2 that makes
creating new partitions increase in time based on the number of
partitions. With thousands of partitions in this import we are spending
~10h waiting for those partitions to create. In a test run with spark3,
which has a new version of hive, these partitions create in only a few
minutes.

This requires a `spark3` connection to be defined within the airflow
admin interface that sets the appropriate spark_home for spark3.

Bug: T326243
Change-Id: I326edf9a82daf413281ddeeb23b07736c11d06e9