cirrus namespaces: Work around missing domain_name in upstream
The canonical_data.wikis table added a column for a new wiki,
azwikimedia, without adding a domain_name. This caused the script to
fail when attempting to fetch from a url with a blank domain name. Make
a bold assumption that we will only see this on brand new wikis and that
it will be fixed with time.

Change-Id: Ib317ccdd070103dd89088f31483a67fac65dcb7e